part2
we are tasked with creating a scanner. this creates tokens out of the
files being compiled.

we retrofitted our main to have the requirements and to output the
correct files. we now need a ".tok" file to be put out.

there is also a dummy parser we are given, we will actually be creating
the parser in the next assignment.