part 3
we are now tasked with creating a parser. we were given a dummy parser
to begin with and an Abtract syntax tree adt.

we also have to retrofit our main.cpp to allow for debug flags to be
used.

we must create another file, with the suffix ".ast"

basically we are trying to to translate that
descriptive user-grammar into LALR(1) form acceptable to the language
bison.