part 4 starting code
we are now tasked with creating a symbol table and type checker for the
compiler. we have to make sure that each case can be checked, and when
necessary, use recursion to check certain cases.

i.e. structs, blocks, etc.

we have to create a new ".sym" file

This is probably the longest part, with the files being probably the
longest.