part 5 (final)
for the final part we have to create an intermediate language to be
emitted by our compiler.

we have to once again, retrofit our main.cpp to output the intermediate
language and a ".oil" file.

this intermediate language is very similar to C.

this part might actually be the longest one, I lied about the last time.

it doesn't work completely, once again it only works for certain cases,
will include test file and test outputs to show what's happening.